Re-architecture Phase 2: extract write-side into ProjectStructureUpdater (seam R1)
Extraction-doc step 8, amended per re-architecture par. 6:

- updateProjectConfiguration, its validators, legacy-dependency upgrading,
  and UpdateBuilder move from ProjectStructure to a standalone
  ProjectStructureUpdater (same package, still in the server module; it is
  destined for L3 in Phase 3 and no longer binds to server infrastructure).
- Seam R1: the three collectUpdateProjectConfigurationOperations calls
  (structure, legacy 4-arg, extension) are folded into one private dispatch
  method, documented as the legacy write contract that the
  layout-reconciliation plan will replace.
- The updater throws LegendSDLCException with int status codes (equal codes
  and messages as before; the server maps them identically via the Phase 2
  mapper). ProjectStructure's serialization trio widens to package-private
  for the updater; protected statics used by structure subclasses stay put.
- No deprecated newUpdateBuilder forwarders are left on ProjectStructure:
  the class leaves the server module next step, so the rename is documented
  in the migration recipe instead. In-repo call sites updated.
