+++ legend-sdlc-server/src/main/java/org/finos/legend/sdlc/server/project/ProjectStructure.java
@@ -26,50 +26,35 @@ import org.eclipse.collections.api.factory.Sets;
 import org.eclipse.collections.api.list.MutableList;
 import org.eclipse.collections.api.map.MutableMap;
 import org.eclipse.collections.api.set.MutableSet;
-import org.eclipse.collections.api.tuple.Pair;
 import org.eclipse.collections.impl.list.mutable.ListAdapter;
-import org.eclipse.collections.impl.tuple.Tuples;
 import org.eclipse.collections.impl.utility.Iterate;
-import org.eclipse.collections.impl.utility.LazyIterate;
-import org.eclipse.collections.impl.utility.ListIterate;
 import org.finos.legend.sdlc.domain.model.entity.Entity;
 import org.finos.legend.sdlc.domain.model.project.ProjectType;
 import org.finos.legend.sdlc.domain.model.project.configuration.ArtifactGeneration;
 import org.finos.legend.sdlc.domain.model.project.configuration.ArtifactType;
 import org.finos.legend.sdlc.domain.model.project.configuration.ArtifactTypeGenerationConfiguration;
 import org.finos.legend.sdlc.domain.model.project.configuration.MetamodelDependency;
-import org.finos.legend.sdlc.domain.model.project.configuration.PlatformConfiguration;
 import org.finos.legend.sdlc.domain.model.project.configuration.ProjectConfiguration;
 import org.finos.legend.sdlc.domain.model.project.configuration.ProjectDependency;
 import org.finos.legend.sdlc.domain.model.project.configuration.ProjectStructureVersion;
 import org.finos.legend.sdlc.domain.model.project.workspace.WorkspaceType;
-import org.finos.legend.sdlc.domain.model.revision.Revision;
 import org.finos.legend.sdlc.domain.model.version.VersionId;
 import org.finos.legend.sdlc.serialization.EntitySerializer;
 import org.finos.legend.sdlc.serialization.EntitySerializers;
-import org.finos.legend.sdlc.server.domain.api.project.ProjectConfigurationUpdater;
-import org.finos.legend.sdlc.server.domain.api.project.source.PatchSourceSpecification;
 import org.finos.legend.sdlc.server.domain.api.project.source.SourceSpecification;
-import org.finos.legend.sdlc.server.domain.api.project.source.SourceSpecificationConsumer;
-import org.finos.legend.sdlc.server.domain.api.project.source.VersionSourceSpecification;
-import org.finos.legend.sdlc.server.domain.api.project.source.WorkspaceSourceSpecification;
 import org.finos.legend.sdlc.server.domain.api.workspace.WorkspaceSpecification;
 import org.finos.legend.sdlc.server.error.LegendSDLCServerException;
 import org.finos.legend.sdlc.server.project.ProjectFileAccessProvider.FileAccessContext;
 import org.finos.legend.sdlc.server.project.ProjectFileAccessProvider.ProjectFile;
 import org.finos.legend.sdlc.server.project.ProjectFileAccessProvider.WorkspaceAccessType;
-import org.finos.legend.sdlc.server.project.extension.ProjectStructureExtension;
-import org.finos.legend.sdlc.server.project.extension.ProjectStructureExtensionProvider;
 import org.finos.legend.sdlc.tools.StringTools;
 
 import javax.lang.model.SourceVersion;
 import javax.ws.rs.core.Response.Status;
 import java.io.InputStream;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Comparator;
-import java.util.EnumSet;
 import java.util.List;
 import java.util.Set;
 import java.util.SortedMap;
@@ -97,8 +82,6 @@ public abstract class ProjectStructure
     private static final Pattern STRICT_VERSION_ID_PATTERN = Pattern.compile("((0|([1-9]\\d*+))\\.){2}(0|([1-9]\\d*+))");
 
     public static final String PROJECT_CONFIG_PATH = "/project.json";
-
-    private static final Set<ArtifactType> FORBIDDEN_ARTIFACT_GENERATION_TYPES = Collections.unmodifiableSet(EnumSet.of(ArtifactType.entities, ArtifactType.versioned_entities, ArtifactType.service_execution));
 
     private final ProjectConfiguration projectConfiguration;
     private final MutableList<EntitySourceDirectory> entitySourceDirectories;
@@ -398,340 +381,11 @@ public abstract class ProjectStructure
         return (versionId != null) && (versionId.length() >= 5) && STRICT_VERSION_ID_PATTERN.matcher(versionId).matches();
     }
 
-    private static Revision updateProjectConfiguration(UpdateBuilder updateBuilder, boolean requireRevisionId)
-    {
-        String projectId = updateBuilder.getProjectId();
-        SourceSpecification sourceSpecification = updateBuilder.getSourceSpecification();
-        ProjectFileAccessProvider projectFileAccessProvider = updateBuilder.getProjectFileAccessProvider();
-
-        String revisionId;
-        if (updateBuilder.getRevisionId() == null)
-        {
-            // if revisionId not specified, get the current revision
-            Revision currentRevision = projectFileAccessProvider.getRevisionAccessContext(projectId, sourceSpecification, null).getCurrentRevision();
-            if (currentRevision != null)
-            {
-                revisionId = currentRevision.getId();
-            }
-            else if (requireRevisionId)
-            {
-                StringBuilder builder = new StringBuilder("Could not find current revision for ");
-                sourceSpecification.visit(new SourceSpecificationConsumer()
-                {
-                    @Override
-                    protected void accept(VersionSourceSpecification sourceSpec)
-                    {
-                        sourceSpec.getVersionId().appendVersionIdString(builder.append("version ")).append(" of ");
-                    }
-
-                    @Override
-                    protected void accept(PatchSourceSpecification sourceSpec)
-                    {
-                        sourceSpec.getVersionId().appendVersionIdString(builder.append("patch version ")).append(" of ");
-                    }
-
-                    @Override
-                    protected void accept(WorkspaceSourceSpecification sourceSpec)
-                    {
-                        WorkspaceSpecification workspaceSpec = sourceSpec.getWorkspaceSpecification();
-                        builder.append(workspaceSpec.getType().getLabel()).append(" ").append(workspaceSpec.getAccessType().getLabel()).append(" ").append(workspaceSpec.getId()).append("in ");
-                    }
-                });
-                builder.append("project ").append(projectId).append(": it may be corrupt");
-                throw new LegendSDLCServerException(builder.toString());
-            }
-            else
-            {
-                revisionId = null;
-            }
-        }
-        else
-        {
-            revisionId = updateBuilder.getRevisionId();
-        }
-
-        FileAccessContext fileAccessContext = CachingFileAccessContext.wrap(projectFileAccessProvider.getFileAccessContext(projectId, sourceSpecification, revisionId));
-
-        ProjectFile configFile = getProjectConfigurationFile(fileAccessContext);
-        ProjectConfiguration currentConfig = (configFile == null) ? getDefaultProjectConfiguration(projectId) : readProjectConfiguration(configFile);
-        ProjectType oldProjectType = currentConfig.getProjectType();
-        // Upgrade old project types
-        if (!ProjectStructure.isValidProjectType(oldProjectType))
-        {
-            oldProjectType = ProjectType.MANAGED;
-            if (updateBuilder.configUpdater.getProjectType() == null)
-            {
-                updateBuilder.configUpdater.setProjectType(oldProjectType);
-            }
-        }
-        ProjectType newProjectType = updateBuilder.configUpdater.getProjectType() == null ? oldProjectType : updateBuilder.configUpdater.getProjectType();
-        // For MANAGED projects
-        if (updateBuilder.projectStructureExtensionProvider != null && newProjectType != ProjectType.EMBEDDED)
-        {
-            // converted from EMBEDDED set version
-            if (oldProjectType == ProjectType.EMBEDDED && updateBuilder.configUpdater.getProjectStructureVersion() == null)
-            {
-                updateBuilder.configUpdater.setProjectStructureVersion(currentConfig.getProjectStructureVersion().getVersion());
-            }
-            // if extension version was not specified, use the latest one for given version
-            if (updateBuilder.configUpdater.getProjectStructureVersion() != null && updateBuilder.configUpdater.getProjectStructureExtensionVersion() == null)
-            {
-                updateBuilder.configUpdater.setProjectStructureExtensionVersion(updateBuilder.projectStructureExtensionProvider.getLatestVersionForProjectStructureVersion(updateBuilder.configUpdater.getProjectStructureVersion()));
-            }
-        }
-        ProjectConfiguration newConfig = updateLegacyDependencies(updateBuilder.getProjectConfigurationUpdater().update(currentConfig), projectFileAccessProvider);
-
-        validateProjectConfiguration(newConfig);
-
-        ProjectStructureVersion oldProjectVersion = currentConfig.getProjectStructureVersion();
-        // prevent extensions on non-managed projects
-        if (newConfig.getProjectType() == ProjectType.EMBEDDED)
-        {
-            if (newConfig.getProjectStructureVersion().getExtensionVersion() != null)
-            {
-                throw new LegendSDLCServerException("Cannot set extensions on project " + projectId + " with " + newConfig.getProjectType() + " type", Status.BAD_REQUEST);
-            }
-            oldProjectVersion = ProjectStructureVersion.newProjectStructureVersion(oldProjectVersion.getVersion());
-        }
-        // prevent downgrading project
-        if (newConfig.getProjectStructureVersion().compareTo(oldProjectVersion) < 0)
-        {
-            throw new LegendSDLCServerException("Cannot change project " + projectId + " from project structure version " + currentConfig.getProjectStructureVersion().toVersionString() + " to version " + newConfig.getProjectStructureVersion().toVersionString(), Status.BAD_REQUEST);
-        }
-
-        // Serialize new configuration and check if it differs from the old
-        byte[] serializedNewConfig = serializeProjectConfiguration(newConfig);
-        if ((configFile != null) && Arrays.equals(serializedNewConfig, configFile.getContentAsBytes()))
-        {
-            // new configuration file is the same as the old
-            return null;
-        }
-
-        List<ProjectFileOperation> operations = Lists.mutable.empty();
-        operations.add((configFile == null) ? ProjectFileOperation.addFile(PROJECT_CONFIG_PATH, serializedNewConfig) : ProjectFileOperation.modifyFile(PROJECT_CONFIG_PATH, serializedNewConfig));
-
-        ProjectStructure currentProjectStructure = getProjectStructure(currentConfig, updateBuilder.getProjectStructurePlatformExtensions());
-        ProjectStructure newProjectStructure = getProjectStructure(newConfig, updateBuilder.getProjectStructurePlatformExtensions());
-
-        // Move or re-serialize entities if necessary
-        List<EntitySourceDirectory> currentEntityDirectories = currentProjectStructure.getEntitySourceDirectories();
-        List<EntitySourceDirectory> newEntityDirectories = newProjectStructure.getEntitySourceDirectories();
-        if (!currentEntityDirectories.equals(newEntityDirectories))
-        {
-            currentEntityDirectories.forEach(currentSourceDirectory -> fileAccessContext.getFilesInDirectory(currentSourceDirectory.getDirectory()).forEach(file ->
-            {
-                String currentPath = file.getPath();
-                if (currentSourceDirectory.isPossiblyEntityFilePath(currentPath))
-                {
-                    byte[] currentBytes = file.getContentAsBytes();
-                    Entity entity;
-                    try
-                    {
-                        entity = currentSourceDirectory.deserialize(currentBytes);
-                    }
-                    catch (Exception e)
-                    {
-                        StringBuilder builder = new StringBuilder("Error deserializing entity from file \"").append(currentPath).append('"');
-                        StringTools.appendThrowableMessageIfPresent(builder, e);
-                        throw new LegendSDLCServerException(builder.toString(), e);
-                    }
-                    EntitySourceDirectory newSourceDirectory = Iterate.detectWith(newEntityDirectories, EntitySourceDirectory::canSerialize, entity);
-                    if (newSourceDirectory == null)
-                    {
-                        throw new LegendSDLCServerException("Could not find a new source directory for entity " + entity.getPath() + ", currently in " + currentPath);
-                    }
-                    if (!currentSourceDirectory.equals(newSourceDirectory))
-                    {
-                        String newPath = newSourceDirectory.entityPathToFilePath(entity.getPath());
-                        byte[] newBytes = newSourceDirectory.serializeToBytes(entity);
-                        if (!newPath.equals(currentPath))
-                        {
-                            operations.add(ProjectFileOperation.moveFile(currentPath, newPath, newBytes));
-                        }
-                        else if (!Arrays.equals(currentBytes, newBytes))
-                        {
-                            operations.add(ProjectFileOperation.modifyFile(currentPath, newBytes));
-                        }
-                    }
-                }
-            }));
-        }
-
-        // Collect any further update operations
-        if (newConfig.getProjectType() != ProjectType.EMBEDDED)
-        {
-            newProjectStructure.collectUpdateProjectConfigurationOperations(currentProjectStructure, fileAccessContext, operations::add);
-        }
-
-        // Call legacy method
-        newProjectStructure.collectUpdateProjectConfigurationOperations(currentProjectStructure, fileAccessContext, (x, y) ->
-        {
-            throw new UnsupportedOperationException();
-        }, operations::add);
-
-        // Collect update operations from any project structure extension
-        if (newConfig.getProjectStructureVersion().getExtensionVersion() != null)
-        {
-            ProjectStructureExtensionProvider extensionProvider = updateBuilder.getProjectStructureExtensionProvider();
-            if (extensionProvider != null)
-            {
-                ProjectStructureExtension projectStructureExtension = extensionProvider.getProjectStructureExtension(newConfig.getProjectStructureVersion().getVersion(), newConfig.getProjectStructureVersion().getExtensionVersion());
-                projectStructureExtension.collectUpdateProjectConfigurationOperations(currentConfig, newConfig, fileAccessContext, operations::add);
-            }
-        }
-
-        // Submit changes
-        return projectFileAccessProvider.getFileModificationContext(projectId, sourceSpecification, revisionId).submit(updateBuilder.getMessage(), operations);
-    }
-
-    private static void validateProjectConfiguration(ProjectConfiguration config)
-    {
-        // Group id
-        if (!isValidGroupId(config.getGroupId()))
-        {
-            throw new LegendSDLCServerException("Invalid groupId: " + config.getGroupId(), Status.BAD_REQUEST);
-        }
-
-        // Artifact id
-        if (!isValidArtifactId(config.getArtifactId()))
-        {
-            throw new LegendSDLCServerException("Invalid artifactId: " + config.getArtifactId() + ". ArtifactId must follow pattern that starts with a lowercase letter and can include lowercase letters, digits, underscores, and hyphens between segments.", Status.BAD_REQUEST);
-        }
-
-        // Project type
-        if (!isValidProjectType(config.getProjectType()))
-        {
-            throw new LegendSDLCServerException("Invalid projectType: " + config.getProjectType(), Status.BAD_REQUEST);
-        }
-
-        // Platform configurations
-        if (config.getPlatformConfigurations() != null)
-        {
-            MutableSet<String> platformNames = Sets.mutable.empty();
-            MutableSet<String> platformNameConflicts = LazyIterate.collect(config.getPlatformConfigurations(), PlatformConfiguration::getName).reject(platformNames::add, Sets.mutable.empty());
-            if (platformNameConflicts.notEmpty())
-            {
-                throw new LegendSDLCServerException(platformNameConflicts.toSortedList().makeString("Platform configuration conflicts: \"", "\", \"", "\""), Status.BAD_REQUEST);
-            }
-        }
-
-        // Metamodel dependencies
-        validateMetamodelDependencies(config.getMetamodelDependencies());
-
-        // Artifact generations
-        validateArtifactGenerations(config.getArtifactGenerations());
-    }
-
-    private static void validateMetamodelDependencies(List<MetamodelDependency> metamodelDependencies)
-    {
-        MutableList<MetamodelDependency> unknownDependencies = ListIterate.reject(metamodelDependencies, ProjectStructure::isKnownMetamodel);
-        if (unknownDependencies.notEmpty())
-        {
-            StringBuilder builder = new StringBuilder("There were issues with one or more added metamodel dependencies");
-            builder.append("; unknown ").append((unknownDependencies.size() == 1) ? "dependency" : "dependencies").append(": ");
-            unknownDependencies.sort(getMetamodelDependencyComparator());
-            unknownDependencies.forEach(d -> d.appendDependencyString((d == unknownDependencies.get(0)) ? builder : builder.append(", ")));
-            throw new LegendSDLCServerException(builder.toString(), Status.BAD_REQUEST);
-        }
-        validateDependencyConflicts(
-                metamodelDependencies,
-                MetamodelDependency::getMetamodel,
-                (id, deps) -> (deps.size() > 1) ? deps.stream().collect(StringBuilder::new, (builder, dep) -> dep.appendVersionIdString(builder.append((builder.length() == 0) ? "multiple versions not allowed: " : ", ")), StringBuilder::append).toString() : null,
-                "metamodels");
-    }
-
-    private static void validateArtifactGenerations(List<ArtifactGeneration> artifactGenerations)
-    {
-        String initString = "There were issues with one or more added artifact generations";
-        StringBuilder builder = null;
-
-        if (Iterate.anySatisfy(artifactGenerations, g -> FORBIDDEN_ARTIFACT_GENERATION_TYPES.contains(g.getType())))
-        {
-            builder = new StringBuilder(initString);
-            LazyIterate.collect(FORBIDDEN_ARTIFACT_GENERATION_TYPES, ArtifactType::getLabel)
-                    .toSortedList()
-                    .appendString(builder, ": generation types ", ", ", " are not allowed");
-        }
-
-        MutableSet<String> artifactGenerationNames = Sets.mutable.empty();
-        MutableSet<String> artifactGenerationConflicts = LazyIterate.collect(artifactGenerations, ArtifactGeneration::getName).reject(artifactGenerationNames::add, Sets.mutable.empty());
-        if (artifactGenerationConflicts.notEmpty())
-        {
-            if (builder == null)
-            {
-                builder = new StringBuilder(initString).append(": ");
-            }
-            else
-            {
-                builder.append("; ");
-            }
-            artifactGenerationConflicts.toSortedList().appendString(builder, "duplicate generations: \"", "\", \"", "\"");
-        }
-
-        if (builder != null)
-        {
-            throw new LegendSDLCServerException(builder.toString(), Status.BAD_REQUEST);
-        }
-    }
-
     protected static boolean isLegacyProjectDependency(ProjectDependency projectDependency)
     {
         return (projectDependency != null) &&
                 (projectDependency.getProjectId() != null) &&
                 (projectDependency.getProjectId().indexOf(':') == -1);
-    }
-
-    private static ProjectConfiguration updateLegacyDependencies(ProjectConfiguration config, ProjectFileAccessProvider projectFileAccessProvider)
-    {
-        MutableList<ProjectDependency> legacyProjectDependencies = Iterate.select(config.getProjectDependencies(), ProjectStructure::isLegacyProjectDependency, Lists.mutable.empty());
-        if (legacyProjectDependencies.isEmpty())
-        {
-            return config;
-        }
-
-        MutableSet<ProjectDependency> projectDependencies = Sets.mutable.withAll(config.getProjectDependencies());
-        MutableList<ProjectDependency> unknownDependencies = Lists.mutable.empty();
-        MutableList<Pair<ProjectDependency, Exception>> accessExceptions = Lists.mutable.empty();
-        legacyProjectDependencies.forEach(dep ->
-        {
-            try
-            {
-                ProjectConfiguration dependencyConfig = getProjectConfiguration(projectFileAccessProvider.getFileAccessContext(dep.getProjectId(), SourceSpecification.versionSourceSpecification(VersionId.parseVersionId(dep.getVersionId()))));
-                if ((dependencyConfig == null) || (dependencyConfig.getArtifactId() == null) || (dependencyConfig.getGroupId() == null))
-                {
-                    unknownDependencies.add(dep);
-                }
-                else
-                {
-                    projectDependencies.remove(dep);
-                    projectDependencies.add(ProjectDependency.newProjectDependency(dependencyConfig.getGroupId() + ":" + dependencyConfig.getArtifactId(), dep.getVersionId()));
-                }
-            }
-            catch (Exception e)
-            {
-                accessExceptions.add(Tuples.pair(dep, e));
-            }
-        });
-        Comparator<ProjectDependency> comparator = getProjectDependencyComparator();
-        if (unknownDependencies.notEmpty() || accessExceptions.notEmpty())
-        {
-            StringBuilder builder = new StringBuilder("There were issues with one or more legacy project dependencies");
-            if (unknownDependencies.notEmpty())
-            {
-                builder.append("; unknown ").append((unknownDependencies.size() == 1) ? "dependency" : "dependencies").append(": ");
-                unknownDependencies.sortThis(comparator).forEach(d -> d.appendDependencyString((d == unknownDependencies.get(0)) ? builder : builder.append(", ")));
-            }
-            if (accessExceptions.notEmpty())
-            {
-                builder.append("; access ").append((accessExceptions.size() == 1) ? "exception" : "exceptions").append(": ");
-                accessExceptions.sortThis(Comparator.comparing(Pair::getOne, comparator)).forEach(p -> p.getOne().appendDependencyString((p == accessExceptions.get(0)) ? builder : builder.append(", ")).append(" (").append(p.getTwo().getMessage()).append(')'));
-            }
-            throw new LegendSDLCServerException(builder.toString(), Status.BAD_REQUEST);
-        }
-
-        SimpleProjectConfiguration newConfig = new SimpleProjectConfiguration(config);
-        newConfig.setProjectDependencies(projectDependencies.toSortedList(comparator));
-        return newConfig;
     }
 
     protected static Comparator<ProjectDependency> getProjectDependencyComparator()
@@ -765,13 +419,12 @@ public abstract class ProjectStructure
         }
     }
 
-
-    private static ProjectFile getProjectConfigurationFile(FileAccessContext accessContext)
+    static ProjectFile getProjectConfigurationFile(FileAccessContext accessContext)
     {
         return accessContext.getFile(PROJECT_CONFIG_PATH);
     }
 
-    private static byte[] serializeProjectConfiguration(ProjectConfiguration projectConfiguration)
+    static byte[] serializeProjectConfiguration(ProjectConfiguration projectConfiguration)
     {
         try
         {
@@ -783,7 +436,7 @@ public abstract class ProjectStructure
         }
     }
 
-    private static ProjectConfiguration readProjectConfiguration(ProjectFile file)
+    static ProjectConfiguration readProjectConfiguration(ProjectFile file)
     {
         try (InputStream stream = file.getContentAsInputStream())
         {
@@ -792,34 +445,6 @@ public abstract class ProjectStructure
         catch (Exception e)
         {
             throw new RuntimeException(StringTools.appendThrowableMessageIfPresent("Error reading project configuration", e), e);
-        }
-    }
-
-    private static boolean isKnownMetamodel(MetamodelDependency metamodelDependency)
-    {
-        // This is a hack until we have a proper metamodel registry
-        String metamodel = metamodelDependency.getMetamodel();
-        int version = metamodelDependency.getVersion();
-        if (metamodel == null)
-        {
-            return false;
-        }
-        switch (metamodel)
-        {
-            case "pure":
-            {
-                return (version == 0) || (version == 1);
-            }
-            case "lineage":
-            case "service":
-            case "tds":
-            {
-                return version == 1;
-            }
-            default:
-            {
-                return false;
-            }
         }
     }
 
@@ -832,213 +457,5 @@ public abstract class ProjectStructure
     protected static EntitySourceDirectory newEntitySourceDirectory(String directory, EntitySerializer serializer)
     {
         return new EntitySourceDirectory(directory, serializer);
-    }
-
-    public static UpdateBuilder newUpdateBuilder(ProjectFileAccessProvider projectFileAccessProvider, String projectId)
-    {
-        return newUpdateBuilder(projectFileAccessProvider, projectId, null);
-    }
-
-    public static UpdateBuilder newUpdateBuilder(ProjectFileAccessProvider projectFileAccessProvider, String projectId, ProjectConfigurationUpdater configUpdater)
-    {
-        return new UpdateBuilder(projectFileAccessProvider, projectId, configUpdater);
-    }
-
-    public static class UpdateBuilder
-    {
-        private final ProjectFileAccessProvider projectFileAccessProvider;
-        private final String projectId;
-        private SourceSpecification sourceSpecification;
-        private ProjectConfigurationUpdater configUpdater;
-        private String revisionId;
-        private String message;
-        private ProjectStructureExtensionProvider projectStructureExtensionProvider;
-        private ProjectStructurePlatformExtensions projectStructurePlatformExtensions;
-
-        private UpdateBuilder(ProjectFileAccessProvider projectFileAccessProvider, String projectId, ProjectConfigurationUpdater configUpdater)
-        {
-            this.projectFileAccessProvider = projectFileAccessProvider;
-            this.projectId = projectId;
-            this.sourceSpecification = SourceSpecification.projectSourceSpecification();
-            this.configUpdater = (configUpdater == null) ? getDefaultProjectConfigurationUpdater() : configUpdater;
-        }
-
-        // Project id
-
-        public String getProjectId()
-        {
-            return this.projectId;
-        }
-
-        public SourceSpecification getSourceSpecification()
-        {
-            return this.sourceSpecification;
-        }
-
-        // Project file access provider
-
-        public ProjectFileAccessProvider getProjectFileAccessProvider()
-        {
-            return this.projectFileAccessProvider;
-        }
-
-        // Project configuration updater
-
-        public ProjectConfigurationUpdater getProjectConfigurationUpdater()
-        {
-            return this.configUpdater;
-        }
-
-        public void setProjectConfigurationUpdater(ProjectConfigurationUpdater updater)
-        {
-            this.configUpdater = (updater == null) ? getDefaultProjectConfigurationUpdater() : updater;
-        }
-
-        public UpdateBuilder withProjectConfigurationUpdater(ProjectConfigurationUpdater updater)
-        {
-            setProjectConfigurationUpdater(updater);
-            return this;
-        }
-
-        private ProjectConfigurationUpdater getDefaultProjectConfigurationUpdater()
-        {
-            return ProjectConfigurationUpdater.newUpdater().withProjectId(this.projectId);
-        }
-
-        // Source specification
-
-        public void setSourceSpecification(SourceSpecification sourceSpec)
-        {
-            this.sourceSpecification = sourceSpec;
-        }
-
-        public UpdateBuilder withSourceSpecification(SourceSpecification sourceSpec)
-        {
-            setSourceSpecification(sourceSpec);
-            return this;
-        }
-
-        public UpdateBuilder withWorkspace(WorkspaceSpecification workspaceSpec)
-        {
-            return withSourceSpecification(SourceSpecification.workspaceSourceSpecification(workspaceSpec));
-        }
-
-        @Deprecated
-        public UpdateBuilder withWorkspace(SourceSpecification sourceSpecification)
-        {
-            return withSourceSpecification(sourceSpecification);
-        }
-
-        // Revision id
-
-        public String getRevisionId()
-        {
-            return this.revisionId;
-        }
-
-        public void setRevisionId(String revisionId)
-        {
-            this.revisionId = revisionId;
-        }
-
-        public UpdateBuilder withRevisionId(String revisionId)
-        {
-            setRevisionId(revisionId);
-            return this;
-        }
-
-        // Message
-
-        public String getMessage()
-        {
-            return this.message;
-        }
-
-        public void setMessage(String message)
-        {
-            this.message = message;
-        }
-
-        public UpdateBuilder withMessage(String message)
-        {
-            setMessage(message);
-            return this;
-        }
-
-        // Project structure extension provider
-
-        public ProjectStructureExtensionProvider getProjectStructureExtensionProvider()
-        {
-            return this.projectStructureExtensionProvider;
-        }
-
-        public void setProjectStructureExtensionProvider(ProjectStructureExtensionProvider projectStructureExtensionProvider)
-        {
-            this.projectStructureExtensionProvider = projectStructureExtensionProvider;
-        }
-
-        public UpdateBuilder withProjectStructureExtensionProvider(ProjectStructureExtensionProvider projectStructureExtensionProvider)
-        {
-            setProjectStructureExtensionProvider(projectStructureExtensionProvider);
-            return this;
-        }
-
-        // Project structure platform extensions
-
-        public ProjectStructurePlatformExtensions getProjectStructurePlatformExtensions()
-        {
-            return this.projectStructurePlatformExtensions;
-        }
-
-        public void setProjectStructurePlatformExtensions(ProjectStructurePlatformExtensions projectStructurePlatformExtensions)
-        {
-            this.projectStructurePlatformExtensions = projectStructurePlatformExtensions;
-        }
-
-        public UpdateBuilder withProjectStructurePlatformExtensions(ProjectStructurePlatformExtensions projectStructurePlatformExtensions)
-        {
-            setProjectStructurePlatformExtensions(projectStructurePlatformExtensions);
-            return this;
-        }
-
-        // Update
-
-        public Revision update()
-        {
-            return update(true);
-        }
-
-        public Revision build()
-        {
-            if (this.configUpdater.getProjectStructureVersion() == null)
-            {
-                this.configUpdater.setProjectStructureVersion(getLatestProjectStructureVersion());
-            }
-            return update(false);
-        }
-
-        private Revision update(boolean requireRevisionId)
-        {
-            if (this.projectId != null)
-            {
-                if (this.configUpdater.getProjectId() == null)
-                {
-                    this.configUpdater.setProjectId(this.projectId);
-                }
-                else if (!this.projectId.equals(this.configUpdater.getProjectId()))
-                {
-                    throw new IllegalArgumentException("Conflicting project ids: \"" + this.projectId + "\" vs \"" + this.configUpdater.getProjectId() + "\"");
-                }
-            }
-            else if (this.configUpdater.getProjectId() == null)
-            {
-                throw new IllegalArgumentException("No project id specified");
-            }
-            if (this.projectStructureExtensionProvider == null && this.configUpdater.getProjectStructureExtensionVersion() != null)
-            {
-                    throw new IllegalArgumentException("Project structure extension version specified (" + this.configUpdater.getProjectStructureExtensionVersion() + ") with no project structure extension provider");
-            }
-            return updateProjectConfiguration(this, requireRevisionId);
-        }
     }
 }
